Change log format
I made a big change to logging. Since I wanted to log out to file, I
added a few code that can handle logging. But error handlings are
still need to refacotred alot. Make it clear when and what causes
rserver to stop to work. I made changes to comamnd options as well.